Enemies now have a sprite
Didn't end up changing a whole lot, but the enemies now look like
zombies and the sprite texture is loaded in main.cpp. There are a couple
bugs where only one enemy animates at a time (I think it's due to the
use of &iterations in the walk method but can't figure what the
alternative could be) and they don't animate when the player is holding
down a direction. I've saved a copy of the version from this morning so
it'd be easy enough to revert it if needed.

+++ Game Start/Effects.cpp
@@ -280,7 +280,7 @@ void Effects::screenShakeUpdate()
     {
 		int xShake = rand() % 10; //gives a random number between 0 and 10 for the X axis
 		int yShake = rand() % 5;
-		_view->move(sf::Vector2f(xShake * power, yShake * power));
+		_view->move(sf::Vector2f((float)xShake * power, (float)yShake * power));
 
         // Increase the shake timer by one frame and check if we have been shaking long enough
         shakeTimer++;
